+++ inc/Runner.php
@@ -54,10 +54,10 @@ class Runner {
 	 * @return bool True on success, false on failure.
 	 */
 	public function delete_local_repository(): bool {
-		/** @var \WP_Filesystem_Base $wp_filesystem */
+		/** @var \WP_Filesystem_Base|null $wp_filesystem */
 		global $wp_filesystem;
 
-		if ( ! $wp_filesystem ) {
+		if ( ! $wp_filesystem instanceof \WP_Filesystem_Base ) {
 			require_once ABSPATH . '/wp-admin/includes/admin.php';
 
 			if ( ! \WP_Filesystem() ) {